+++ example/SubApp/React/webpack.config.js
@@ -9,8 +9,13 @@ chainOsWebpack({
   id: 'os-example'
 })(chain);
 
+const regCommon = new RegExp(`[\\/]node_modules[\\/](_)?(${['react'].join('|')})[\\/|@]`);
+
 module.exports = merge(chain.toConfig(), {
-  entry: './src/index.js',
+  entry: {
+    'os-example': './src/index.js',
+    'os-example2': './src/index2.js'
+  },
   devtool: 'source-map',
   devServer: {
     port: '8081',
@@ -44,6 +49,17 @@ module.exports = merge(chain.toConfig(), {
         use: ['style-loader', 'css-loader', 'less-loader'],
       },
     ],
+  },
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        commons: {
+          test: regCommon,
+          name: 'common',
+          chunks: 'all'
+        }
+      }
+    }
   },
   plugins: [
     new HtmlWebpackPlugin({