+++ packages/build-tools/breezr-plugin-os/src/index.ts
@@ -2,8 +2,9 @@ import * as WebpackChain from 'webpack-chain';
 import { PluginAPI, PluginOptions } from '@alicloud/console-toolkit-core';
 import { OSJsonpWebpackPlugin } from './OSJsonpPlugin';
 import { DonePlugin } from './DonePlugins';
+import { MultiEntryManifest } from './MultiEntryManifest';
 import * as WebpackAssetsManifestPlugin from 'webpack-assets-manifest';
-import { sandBoxCss } from './sandboxCss';
+import { wrapCss } from 'postcss-prefix-wrapper';
 
 export const chainOsWebpack = (options: PluginOptions) => async (config: WebpackChain) => {
   const { jsonpCall, injectVars } = options;
@@ -39,10 +40,18 @@ export const chainOsWebpack = (options: PluginOptions) => async (config: Webpack
     entrypoints: true,
     output: `${options.id}.manifest.json`
   }]);
+
+  config.plugin('MultiEntryManifest').use(MultiEntryManifest, [{
+    entryName: `${options.id}.manifest.json`,
+  }]);
   
   config.plugin('WebpackDonePlugin').use(DonePlugin, [{
     done: () => {
-      sandBoxCss(options.cssBuildDir || config.output.get('path'), options.id, options)
+      // process css
+      wrapCss(options.cssBuildDir || config.output.get('path'), options.id, {
+        ext: '.os.css',
+        disableOsCssExtends: options.disableOsCssExtends,
+      })
     }
   }])
 }