+++ packages/build-tools/postcss-wrapper/src/index.ts
@@ -1,0 +1,4 @@
+import { postcssWrap } from './postcssWrap';
+export { wrapCss } from './wrapCss';
+
+export default postcssWrap;