+++ packages/core/kernal/src/type.ts
@@ -25,6 +25,12 @@ export interface AppInfo<T = any> extends BasicModule, ExtendsAppLifeCycles<T> {
   version?: string;
   name?: string;
   dom?: Element;
+  logger?: {
+    debug: () => {};
+    error: () => {};
+    warn: () => {};
+    info: () => {};
+  };
   manifest?: string;
   externals?: BasicModule[];
   customProps?: {
@@ -46,6 +52,7 @@ export interface SandBoxOption {
   initialPath?: string;
   externalsVars?: string[];
   allowEvents?: string[];
+  disableFakeBody?: boolean;
 }
 
 export interface AppOption {